// src/utils/enums/affine-block-flavour.enum.ts
export enum AffineBlockFlavour {
  PAGE = 'affine:page',
  SURFACE = 'affine:surface',
  NOTE = 'affine:note',
  PARAGRAPH = 'affine:paragraph',
  LIST = 'affine:list',
  CODE = 'affine:code',
  DIVIDER = 'affine:divider',
  BOOKMARK = 'affine:bookmark',
  EMBED_LINKED_DOC = 'affine:embed-linked-doc',
  EMBED_SYNCED_DOC = 'affine:embed-synced-doc',
  IMAGE = 'affine:image',
  LATEX = 'affine:latex',
  TABLE = 'affine:table',
  TABLE_ROW = 'affine:table-row',
  TABLE_CELL = 'affine:table-cell',
  CALLOUT = 'affine:callout',
  ATTACHMENT = 'affine:attachment',
}